% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/activeSpace.R
\name{activeSpace}
\alias{activeSpace}
\title{Active Space Formula}
\usage{
activeSpace(SL, NL, freq, r = c(0, 1000), PL = 20, boolR = T)
}
\arguments{
\item{SL}{Source Level measured from the field}

\item{NL}{Noise Level gathered from wenz graph (specLvlGraph)}

\item{freq}{The  frequency of center frequency used in calculating the absorbance}

\item{r}{A vector of length two where the radius of our active space may lie}

\item{PL}{The propogation loss (default is 20). Gathered from measurements.}

\item{boolR}{Boolean of whether you want the value printed out in a string. Should be true for HW problems.}
}
\value{
The active space in meters
}
\description{
This function calculates the active space (AS)
}
