% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addDB.R
\name{addDB}
\alias{addDB}
\title{Add dBs}
\usage{
addDB(dbs, PL = 20)
}
\arguments{
\item{dbs}{A vector of dB values}

\item{PL}{The reference value where pressure dB is 20 and intensity dB is 10}
}
\value{
The addition of a series of dB levels
}
\description{
This function adds dB levels
}
