% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/avgDB.R
\name{avgDB}
\alias{avgDB}
\title{Average dBs}
\usage{
avgDB(dbs, PL = 20)
}
\arguments{
\item{dbs}{A vector of dB values}

\item{PL}{The reference value where pressure dB is 20 and intensity dB is 10}
}
\value{
The average of a series of dB levels
}
\description{
This function averages dB levels
}
