% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/avgSL.R
\name{avgSL}
\alias{avgSL}
\title{Average Source Level}
\usage{
avgSL(press, changingR = F, ref = 20 * 10^(-6), r = 50, PL = 20, boolR = T)
}
\arguments{
\item{press}{Measured pressure values (prms, Pa)
Source-receiver distance (m)
@param r if it is a constant
@param changingR Boolean for if the distance in which we are measuring the sound changes}

\item{ref}{The pressure reference value e.g. 20*10^(-6) in the air, 10^(-6) underwater.}

\item{PL}{The propogation loss measured for the frequency of the call at the location of the recording system}

\item{boolR}{Boolean to print or the value result as a string. Should be true for HW problems.}
}
\value{
The average source level in dB re. ref Pa.
}
\description{
This function calculates the average source level given a series of pressures measured at r distance from the source
}
