% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opDB.R
\name{opDB}
\alias{opDB}
\title{opDB}
\usage{
opDB(dbs, PL, op = "add")
}
\arguments{
\item{dbs}{A vector of dB values of all the same reference.}

\item{PL}{Should be 10 if dB measurements were for intensity or power and 20 if dB measurements were for pressure.}

\item{op}{Default = 'add'. Can be 'add', 'sub', or 'avg' to add, subtract, or average, respectively.}
}
\value{
A sinlge dB value
}
\description{
This function either adds, subtracts or averages dB values specified in function. This is for dB values in re to power, intensity or pressure.
}
\examples{
#Given a set of intensity values, add them.
opDB(c(100, 101, 127, 96), 10)
}
