% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recievedL2.R
\name{recievedL2}
\alias{recievedL2}
\title{Received Level}
\usage{
recievedL2(SL, r, PL = 20)
}
\arguments{
\item{SL}{Source Level measured from the field}

\item{r}{Distance of call}

\item{PL}{The propogation loss (default is 20). Gathered from measurements.}
}
\value{
The received level in dB relative to pressure reference
}
\description{
This function calculates the received level between two individuals at a particular distance.
}
