% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/activeSpace.R
\name{recievedLevel}
\alias{recievedLevel}
\title{Received Level Function}
\usage{
recievedLevel(r, SL, NL, freq, PL)
}
\arguments{
\item{r}{A vector of length two where the radius of our active space may lie}

\item{SL}{Source Level measured from the field}

\item{NL}{Noise Level gathered from wenz graph (specLvlGraph)}

\item{freq}{The  frequency of center frequency used in calculating the absorbance}

\item{PL}{The propogation loss (default is 20). Gathered from measurements.}
}
\value{
The received level
}
\description{
This function calculates the recieved level of a source level given noise level and absorption
}
