% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmax.R
\name{rmax}
\alias{rmax}
\title{This function solves the sonar equation for the maximum range, for cylindrical spreading and no absorption
e.g seek the range at wich the TL =  SL - NL - DT;}
\usage{
rmax(TL, boolR = T)
}
\description{
This function solves the sonar equation for the maximum range, for cylindrical spreading and no absorption
e.g seek the range at wich the TL =  SL - NL - DT;
}
