% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soundSpeed.R
\name{soundSpeed}
\alias{soundSpeed}
\title{Sound Speed}
\usage{
soundSpeed(t, med = "air", s = NULL, z = NULL)
}
\arguments{
\item{t}{The temperature in °C}

\item{med}{The medium sound is traveling through}

\item{s}{The salinity in parts per thousand (ppt)}

\item{z}{The depth in m}
}
\value{
The speed of sound in m/s
}
\description{
This function finds the average speed of sound in water or air (c) in m/s
given information on temperature, salinity, and depth.
}
\examples{
#Speed of sound in 30°C water with a salinity of 2000 ppt and a depth of
# 2010 m
soundSpeed(30, "water", 2000, 2010)

}
