% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subDB.R
\name{sub}
\alias{sub}
\title{Recursive Subtraction}
\usage{
sub(x)
}
\arguments{
\item{x}{A series of values to subtract}
}
\value{
The solution of subtracting all values
}
\description{
This function subtracts dB levels
}
