% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subDB.R
\name{subDB}
\alias{subDB}
\title{Subtract dBs}
\usage{
subDB(dbs, PL = 20)
}
\arguments{
\item{dbs}{A vector of dB values}

\item{PL}{The reference value where pressure dB is 20 and intensity dB is 10}
}
\value{
The subtraction of a series of dB levels
}
\description{
This function subtracts dB levels
}
