% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/avgSL.R
\name{avgSL}
\alias{avgSL}
\title{Average Source Level}
\usage{
avgSL(press, changingR = F, ref = 20 * 10^(-6), r = 50, PL = 20, boolR = T)
}
\arguments{
\item{press}{Measured pressure values}

\item{changingR}{Boolean for if the distance in which we are measuring the sound change}

\item{ref}{The reference value}

\item{r}{The distance of the source if constant}

\item{PL}{The propogation loss measured from our instrument}

\item{boolR}{Boolean of whether you want the value printed out in a string. Should be true for HW problems.}
}
\value{
The average source level in dB rel to pressure references
}
\description{
This function calculates the average source level given a series of pressures
}
