% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snell.R
\name{snell}
\alias{snell}
\title{Snell's Law}
\usage{
snell(ang, v1, v2)
}
\arguments{
\item{ang}{The known angle of relfection or refraction}

\item{v1}{The longitudinal wave velocity of medium where angle (refraction or reflection) is known given in m/s.}

\item{v2}{The longitudinal wave velocity of medium where angle (refraction or reflection) is unknown given in m/s.}
}
\value{
The opposing angle of reflection or refraction, respectfully
}
\description{
This function is a reformated version of Snell's law that finds reflection or refraction angle given the two mediums longitudinal
wave velocities.
}
\examples{

}
