% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wenz.r
\name{specLvlGraph}
\alias{specLvlGraph}
\title{Soectrum Level Graph and Marine Noise Level}
\usage{
specLvlGraph(freqBand, shipT = -1, seaState = -1, wSpeed = 0, boolR = T)
}
\arguments{
\item{freqBand}{The band of frequencies for an animals call}

\item{shipT}{The ship traffic in the area}

\item{seaState}{The sea state as specified by the National Weather Service}

\item{wSpeed}{The wind speed in miles per hour (mph)}

\item{boolR}{Boolean of whether you want the value printed out in a string. Should be true for HW problems.}
}
\value{
The noise level (NL) in dB rel 1 microP
}
\description{
This function calculates the noise level (NL) in a specific marine environment.
}
